+++ samples/HelloWorld.UI/ViewModels/HelloWorldViewModel.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 
-namespace HelloWorld.ViewModels;
+namespace HelloWorld.UI.ViewModels;
 
 public sealed partial class HelloWorldViewModel : ObservableObject {