+++ samples/HelloWorld/AvaloniaLoader.cs
@@ -1,5 +1,6 @@
 ﻿using Avalonia;
 using Godot;
+using HelloWorld.UI;
 using JLeb.Estragonia;
 
 namespace HelloWorld;