+++ templates/estragonia-godot/GodotGame/AvaloniaLoader.cs
@@ -1,5 +1,6 @@
 using Avalonia;
 using Godot;
+using GodotGame.UI;
 using JLeb.Estragonia;
 
 namespace GodotGame;